real_time_payments/requirements.txt
kafka-python==2.0.2
pymongo==4.5.0
python-dotenv==1.0.0
six==1.16.0
confluent-kafka==2.2.0